requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==3.3.3
django-timezone-field==6.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.5
kombu==5.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
xlrd==2.0.1
xlwt==1.3.0
